MODE-1419 Enabled full text searching
Added full-text search back in. Note that it can be explicitly disabled/enabled within a repository configuration,
and it is currently *enabled* by default. Quite a few test cases were added back in, and some of them highlighted
some issues with search/query scores (were floats, but expected to be doubles).

Also, one quirk of the AS7 subsystem startup is that the text extractors are added to a running repository,
and that means that there's a time during startup when there are no text extractors (and any added binary
values don't get extracted during storage), followed quickly by the enabling of text extractors (by the time
the property with the binary value is indexed). This manifested itself as a blocked sequencing thread within
the ZIP Sequencer integration test (using AS7).

All unit and integration tests pass with these changes, including those that were enabled by these changes.

+++ modeshape-jcr/src/main/java/org/modeshape/jcr/query/XPathQueryResult.java
@@ -132,7 +132,7 @@ public class XPathQueryResult extends JcrQueryResult {
                 return ((XPathQueryResultRowIterator)iterator).jcrPath(location.getPath());
             }
             if (JCR_SCORE_COLUMN_NAME.equals(columnName)) {
-                Float score = (Float)tuple[iterator.scoreIndex];
+                Double score = (Double)tuple[iterator.scoreIndex];
                 return ((XPathQueryResultRowIterator)iterator).jcrDouble(score);
             }
             return super.getValue(columnName);
