+++ modeshape-jcr/src/main/java/org/modeshape/jcr/query/lucene/basic/BasicTupleCollector.java
@@ -293,7 +293,7 @@ public class BasicTupleCollector extends TupleCollector {
         @Override
         protected Object computeValue( Path path,
                                        float score ) {
-            return new Float(score);
+            return new Double(score);
         }
     }
